requirements.txt
astroid==2.3.3
Click==7.0
colorama==0.4.3
dnspython==1.16.0
Flask==1.1.1
Flask-MySQL==1.5.1
Flask-MySQLdb==0.2.0
Flask-WTF==0.14.3
gunicorn==20.0.4
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.11.1
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
mysql-connector==2.2.9
mysql-connector-python==8.0.19
mysqlclient==1.4.6
passlib==1.7.2
protobuf==3.6.1
pylint==2.4.4
PyMySQL==0.9.3
six==1.14.0
Werkzeug==1.0.0
wrapt==1.11.2
WTForms==2.2.1
